// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  ADMIN: 'ADMIN',
  TECH: 'TECH'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const ServiceOrderType = {
  instalacao: 'instalacao',
  manutencao: 'manutencao',
  vistoria: 'vistoria',
  suporte: 'suporte'
} as const

export type ServiceOrderType = (typeof ServiceOrderType)[keyof typeof ServiceOrderType]


export const ServiceOrderDeadline = {
  sem_prazo: 'sem_prazo',
  D1_dia: 'D1_dia',
  D3_dias: 'D3_dias',
  D7_dias: 'D7_dias',
  D15_dias: 'D15_dias',
  D30_dias: 'D30_dias'
} as const

export type ServiceOrderDeadline = (typeof ServiceOrderDeadline)[keyof typeof ServiceOrderDeadline]


export const ServiceOrderStatus = {
  OPEN: 'OPEN',
  IN_PROGRESS: 'IN_PROGRESS',
  DONE: 'DONE',
  CANCELED: 'CANCELED'
} as const

export type ServiceOrderStatus = (typeof ServiceOrderStatus)[keyof typeof ServiceOrderStatus]